+++ packages/core/src/server/storage/types.ts
@@ -3,6 +3,7 @@ import { TestRunValueInfo } from '../../shared/value-info';
 export type ITestRunStorage = {
   load(key: string): Promise<TestRunValueInfo | undefined>;
   save(valueInfo: TestRunValueInfo, options?: { notify?: boolean }): Promise<void>;
+  delete(key: string): Promise<void>;
   waitForComputed(key: string): Promise<TestRunValueInfo>;
   claimForComputing(valueInfo: TestRunValueInfo): boolean;
 };